+++ android/src/main/java/vn/hunghd/flutterdownloader/DownloadWorker.java
@@ -72,6 +72,7 @@ public class DownloadWorker extends Worker implements MethodChannel.MethodCallHa
     public static final String ARG_NOTIFICATION_TITLE = "notification_title";
     public static final String ARG_CALLBACK_HANDLE = "callback_handle";
     public static final String ARG_DEBUG = "debug";
+    public static final String ARG_SAVE_IN_PUBLIC_STORAGE = "save_in_public_storage";
 
     private static final String TAG = DownloadWorker.class.getSimpleName();
     private static final int BUFFER_SIZE = 4096;
@@ -96,6 +97,7 @@ public class DownloadWorker extends Worker implements MethodChannel.MethodCallHa
     private int primaryId;
     private String msgStarted, msgInProgress, msgCanceled, msgFailed, msgPaused, msgComplete;
     private long lastCallUpdateNotification = 0;
+    private boolean saveInPublicStorage;
 
     public DownloadWorker(@NonNull final Context context,
                           @NonNull WorkerParameters params) {
@@ -168,7 +170,6 @@ public class DownloadWorker extends Worker implements MethodChannel.MethodCallHa
         }
     }
 
-
     @NonNull
     @Override
     public Result doWork() {
@@ -193,7 +194,8 @@ public class DownloadWorker extends Worker implements MethodChannel.MethodCallHa
 
         DownloadTask task = taskDao.loadTask(getId().toString());
 
-        log("DownloadWorker{url=" + url + ",filename=" + filename + ",savedDir=" + savedDir + ",header=" + headers + ",isResume=" + isResume);
+        log("DownloadWorker{url=" + url + ",filename=" + filename + ",savedDir=" + savedDir + ",header=" + headers + ",isResume=" + isResume + ",status=" + (task != null ? task.status : "GONE"));
+
 
         // Task has been deleted or cancelled
         if (task == null || task.status == DownloadStatus.CANCELED) {
@@ -203,6 +205,7 @@ public class DownloadWorker extends Worker implements MethodChannel.MethodCallHa
         showNotification = getInputData().getBoolean(ARG_SHOW_NOTIFICATION, false);
         clickToOpenDownloadedFile = getInputData().getBoolean(ARG_OPEN_FILE_FROM_NOTIFICATION, false);
         String notificationTitle = getInputData().getString(ARG_NOTIFICATION_TITLE);
+        saveInPublicStorage = getInputData().getBoolean(ARG_SAVE_IN_PUBLIC_STORAGE, false);
 
         primaryId = task.primaryId;
 
@@ -258,6 +261,7 @@ public class DownloadWorker extends Worker implements MethodChannel.MethodCallHa
         conn.setRequestProperty("Accept-Encoding", "identity");
         conn.setRequestProperty("Range", "bytes=" + downloadedBytes + "-");
         conn.setDoInput(true);
+
         return downloadedBytes;
     }
 
@@ -267,8 +271,8 @@ public class DownloadWorker extends Worker implements MethodChannel.MethodCallHa
         Map<String, Integer> visited;
         HttpURLConnection httpConn = null;
         InputStream inputStream = null;
-        FileOutputStream outputStream = null;
         String saveFilePath;
+        OutputStream outputStream = null;
         String location;
         long downloadedBytes = 0;
         int responseCode;
@@ -310,6 +314,8 @@ public class DownloadWorker extends Worker implements MethodChannel.MethodCallHa
                     case HttpURLConnection.HTTP_MOVED_PERM:
                     case HttpURLConnection.HTTP_SEE_OTHER:
                     case HttpURLConnection.HTTP_MOVED_TEMP:
+                    case 307: /* HTTP_TEMP_REDIRECT */
+                    case 308: /* HTTP_PERM_REDIRECT */
                         log("Response with redirection code");
                         location = httpConn.getHeaderField("Location");
                         log("Location = " + location);
@@ -324,9 +330,9 @@ public class DownloadWorker extends Worker implements MethodChannel.MethodCallHa
             }
 
             httpConn.connect();
-
+            String contentType;
             if ((responseCode == HttpURLConnection.HTTP_OK || (isResume && responseCode == HttpURLConnection.HTTP_PARTIAL)) && !isStopped()) {
-                String contentType = httpConn.getContentType();
+                contentType = httpConn.getContentType();
                 long contentLength = android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.N ? httpConn.getContentLengthLong() : httpConn.getContentLength();
                 log("Content-Type = " + contentType);
                 log("Content-Length = " + contentLength);
@@ -351,7 +357,6 @@ public class DownloadWorker extends Worker implements MethodChannel.MethodCallHa
                         }
                     }
                 }
-                saveFilePath = savedDir + File.separator + filename;
 
                 taskDao.updateTask(getId().toString(), filename, contentType);
 
@@ -359,9 +364,35 @@ public class DownloadWorker extends Worker implements MethodChannel.MethodCallHa
                 inputStream = httpConn.getInputStream();
                 outputStream = new FileOutputStream(saveFilePath, isResume);
 
+
+                String savedFilePath;
+                // opens an output stream to save into file
+                // there are two case:
+                if (isResume) {
+                    // 1. continue downloading (append data to partial downloaded file)
+                    savedFilePath = savedDir + File.separator + filename;
+                    outputStream = new FileOutputStream(savedFilePath, true);
+                } else {
+                    // 2. new download, create new file
+                    // there are two case according to Android SDK version and save path
+                    // From Android 11 onwards, file is only downloaded to app-specific directory (internal storage)
+                    // or public shared download directory (external storage).
+                    // The second option will ignore `savedDir` parameter.
+                    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q && saveInPublicStorage) {
+                        Uri uri = createFileInPublicDownloadsDir(filename, contentType);
+                        savedFilePath = getMediaStoreEntryPathApi29(uri);
+                        outputStream = context.getContentResolver().openOutputStream(uri, "w");
+                    } else {
+                        File file = createFileInAppSpecificDir(filename, savedDir);
+                        savedFilePath = file.getPath();
+                        outputStream = new FileOutputStream(file, false);
+                    }
+                }
+
                 long count = downloadedBytes;
                 int bytesRead;
                 byte[] buffer = new byte[BUFFER_SIZE];
+                // using isStopped() to monitor canceling task
                 while ((bytesRead = inputStream.read(buffer)) != -1 && !isStopped()) {
                     count += bytesRead;
                     int progress = (int) ((count * 100) / (contentLength + downloadedBytes));
@@ -371,14 +402,13 @@ public class DownloadWorker extends Worker implements MethodChannel.MethodCallHa
                             && progress != lastProgress) {
                         lastProgress = progress;
 
-                        updateNotification(context, filename, DownloadStatus.RUNNING, progress, null, false, notificationTitle);
-
                         // This line possibly causes system overloaded because of accessing to DB too many ?!!!
                         // but commenting this line causes tasks loaded from DB missing current downloading progress,
                         // however, this missing data should be temporary and it will be updated as soon as
                         // a new bunch of data fetched and a notification sent
                         taskDao.updateTask(getId().toString(), DownloadStatus.RUNNING, progress);
 
+                        updateNotification(context, filename, DownloadStatus.RUNNING, progress, null, false, notificationTitle);
                     }
                 }
 
@@ -388,22 +418,46 @@ public class DownloadWorker extends Worker implements MethodChannel.MethodCallHa
                 PendingIntent pendingIntent = null;
 
                 updateNotification(context, filename, status, progress, pendingIntent, true, notificationTitle);
+
+                if (status == DownloadStatus.COMPLETE) {
+                    if (Build.VERSION.SDK_INT < Build.VERSION_CODES.Q) {
+                        if (isImageOrVideoFile(contentType) && isExternalStoragePath(savedFilePath)) {
+                            addImageOrVideoToGallery(filename, savedFilePath, getContentTypeWithoutCharset(contentType));
+                        }
+                    }
+
+                    if (clickToOpenDownloadedFile) {
+                        if (android.os.Build.VERSION.SDK_INT < Build.VERSION_CODES.Q && storage != PackageManager.PERMISSION_GRANTED)
+                            return;
+                        Intent intent = IntentUtils.validatedFileIntent(getApplicationContext(), savedFilePath, contentType);
+                        if (intent != null) {
+                            log("Setting an intent to open the file " + savedFilePath);
+                            int flags = Build.VERSION.SDK_INT >= Build.VERSION_CODES.S ? PendingIntent.FLAG_CANCEL_CURRENT | PendingIntent.FLAG_IMMUTABLE : PendingIntent.FLAG_CANCEL_CURRENT;
+                            pendingIntent = PendingIntent.getActivity(getApplicationContext(), 0, intent, flags);
+                        } else {
+                            log("There's no application that can open the file " + savedFilePath);
+                        }
+                    }
+                }
                 taskDao.updateTask(getId().toString(), status, progress);
+                updateNotification(context, filename, status, progress, pendingIntent, true, notificationTitle);
+
 
                 log(isStopped() ? "Download canceled" : "File downloaded");
             } else {
                 DownloadTask task = taskDao.loadTask(getId().toString());
                 int status = isStopped() ? (task.resumable ? DownloadStatus.PAUSED : DownloadStatus.CANCELED) : DownloadStatus.FAILED;
-                updateNotification(context, filename == null ? fileURL : filename, status, -1, null, true, notificationTitle);
                 taskDao.updateTask(getId().toString(), status, lastProgress);
+                updateNotification(context, filename == null ? fileURL : filename, status, -1, null, true, notificationTitle);
                 log(isStopped() ? "Download canceled" : "Server replied HTTP code: " + responseCode);
             }
         } catch (IOException e) {
-            updateNotification(context, filename == null ? fileURL : filename, DownloadStatus.FAILED, -1, null, true, notificationTitle);
             taskDao.updateTask(getId().toString(), DownloadStatus.FAILED, lastProgress);
+            updateNotification(context, filename == null ? fileURL : filename, DownloadStatus.FAILED, -1, null, true, notificationTitle);
             e.printStackTrace();
         } finally {
             if (outputStream != null) {
+                outputStream.flush();
                 try {
                     outputStream.close();
                 } catch (IOException e) {
@@ -419,6 +473,94 @@ public class DownloadWorker extends Worker implements MethodChannel.MethodCallHa
             }
             if (httpConn != null) {
                 httpConn.disconnect();
+            }
+        }
+    }
+
+
+    /**
+     * Create a file using java.io API
+     */
+    private File createFileInAppSpecificDir(String filename, String savedDir) {
+        File newFile = new File(savedDir, filename);
+        try {
+            boolean rs = newFile.createNewFile();
+            if (rs) {
+                return newFile;
+            } else {
+                logError("It looks like you are trying to save file in public storage but not setting 'saveInPublicStorage' to 'true'");
+            }
+        } catch (IOException e) {
+            e.printStackTrace();
+            logError("Create a file using java.io API failed ");
+        }
+        return null;
+    }
+
+    /**
+     * Create a file inside the Download folder using MediaStore API
+     */
+    @RequiresApi(Build.VERSION_CODES.Q)
+    private Uri createFileInPublicDownloadsDir(String filename, String mimeType) {
+        Uri collection = MediaStore.Downloads.EXTERNAL_CONTENT_URI;
+        ContentValues values = new ContentValues();
+        values.put(MediaStore.Downloads.DISPLAY_NAME, filename);
+        values.put(MediaStore.Downloads.MIME_TYPE, mimeType);
+        values.put(MediaStore.Downloads.RELATIVE_PATH, Environment.DIRECTORY_DOWNLOADS);
+        ContentResolver contentResolver = getApplicationContext().getContentResolver();
+        try {
+            return contentResolver.insert(collection, values);
+        } catch (Exception e) {
+            e.printStackTrace();
+            logError("Create a file using MediaStore API failed.");
+        }
+        return null;
+    }
+
+    /**
+     * Get a path for a MediaStore entry as it's needed when calling MediaScanner
+     */
+    private String getMediaStoreEntryPathApi29(Uri uri) {
+        try (Cursor cursor = getApplicationContext().getContentResolver().query(
+                uri,
+                new String[]{MediaStore.Files.FileColumns.DATA},
+                null,
+                null,
+                null
+        )) {
+            if (cursor == null)
+                return null;
+            if (!cursor.moveToFirst())
+                return null;
+            return cursor.getString(cursor.getColumnIndexOrThrow(MediaStore.Files.FileColumns.DATA));
+        } catch (IllegalArgumentException e) {
+            e.printStackTrace();
+            logError("Get a path for a MediaStore failed");
+            return null;
+        }
+    }
+
+    void scanFilePath(String path, String mimeType, CallbackUri callback) {
+        MediaScannerConnection.scanFile(
+                getApplicationContext(),
+                new String[]{path},
+                new String[]{mimeType},
+                (path1, uri) -> callback.invoke(uri));
+    }
+
+    private void cleanUp() {
+        DownloadTask task = taskDao.loadTask(getId().toString());
+        if (task != null && task.status != DownloadStatus.COMPLETE && !task.resumable) {
+            String filename = task.filename;
+            if (filename == null) {
+                filename = task.url.substring(task.url.lastIndexOf("/") + 1, task.url.length());
+            }
+
+            // check and delete uncompleted file
+            String saveFilePath = task.savedDir + File.separator + filename;
+            File tempFile = new File(saveFilePath);
+            if (tempFile.exists()) {
+                tempFile.delete();
             }
         }
     }
@@ -610,9 +752,71 @@ public class DownloadWorker extends Worker implements MethodChannel.MethodCallHa
         return URLDecoder.decode(name, charset != null ? charset : "ISO-8859-1");
     }
 
+    private String getContentTypeWithoutCharset(String contentType) {
+        if (contentType == null)
+            return null;
+        return contentType.split(";")[0].trim();
+    }
+
+    private boolean isImageOrVideoFile(String contentType) {
+        contentType = getContentTypeWithoutCharset(contentType);
+        return (contentType != null && (contentType.startsWith("image/") || contentType.startsWith("video")));
+    }
+
+    private boolean isExternalStoragePath(String filePath) {
+        File externalStorageDir = Environment.getExternalStorageDirectory();
+        return filePath != null && externalStorageDir != null && filePath.startsWith(externalStorageDir.getPath());
+    }
+
+    private void addImageOrVideoToGallery(String fileName, String filePath, String contentType) {
+        if (contentType != null && filePath != null && fileName != null) {
+            if (contentType.startsWith("image/")) {
+                ContentValues values = new ContentValues();
+
+                values.put(MediaStore.Images.Media.TITLE, fileName);
+                values.put(MediaStore.Images.Media.DISPLAY_NAME, fileName);
+                values.put(MediaStore.Images.Media.DESCRIPTION, "");
+                values.put(MediaStore.Images.Media.MIME_TYPE, contentType);
+                values.put(MediaStore.Images.Media.DATE_ADDED, System.currentTimeMillis());
+                values.put(MediaStore.Images.Media.DATE_TAKEN, System.currentTimeMillis());
+                values.put(MediaStore.Images.Media.DATA, filePath);
+
+                log("insert " + values + " to MediaStore");
+
+                ContentResolver contentResolver = getApplicationContext().getContentResolver();
+                contentResolver.insert(MediaStore.Images.Media.EXTERNAL_CONTENT_URI, values);
+            } else if (contentType.startsWith("video")) {
+                ContentValues values = new ContentValues();
+
+                values.put(MediaStore.Video.Media.TITLE, fileName);
+                values.put(MediaStore.Video.Media.DISPLAY_NAME, fileName);
+                values.put(MediaStore.Video.Media.DESCRIPTION, "");
+                values.put(MediaStore.Video.Media.MIME_TYPE, contentType);
+                values.put(MediaStore.Video.Media.DATE_ADDED, System.currentTimeMillis());
+                values.put(MediaStore.Video.Media.DATE_TAKEN, System.currentTimeMillis());
+                values.put(MediaStore.Video.Media.DATA, filePath);
+
+                log("insert " + values + " to MediaStore");
+
+                ContentResolver contentResolver = getApplicationContext().getContentResolver();
+                contentResolver.insert(MediaStore.Video.Media.EXTERNAL_CONTENT_URI, values);
+            }
+        }
+    }
+
     private void log(String message) {
         if (debug) {
             Log.d(TAG, message);
         }
+    }
+
+    private void logError(String message) {
+        if (debug) {
+            Log.e(TAG, message);
+        }
+    }
+
+    public interface CallbackUri {
+        void invoke(Uri uri);
     }
 }